Feat/rename to repoverse (#1)
* Refactor README.md formatting

* Refactor code structure and rename RepoSyncService to Repoverse

* Update release workflow

* Add push trigger for main branch in release workflow

* Update GitHub Actions workflow for PR builds

* Update GitHub Actions workflows

This commit updates the GitHub Actions workflows for pull requests and releases. It includes changes to the steps for checking out the repository, setting up the Node.js environment, installing dependencies, building the project, and running the test suite. The changes ensure that the workflows are up to date and aligned with best practices.

* Update package.json version to 0.0.2-development